+++ src/Confluent.Kafka/Impl/NativeMethods/NativeMethods_Debian9.cs
@@ -347,6 +347,20 @@ namespace Confluent.Kafka.Impl.NativeMethods
                 /* const rd_kafka_topic_partition_list_t * */ IntPtr partitions);
 
         [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
+        internal static extern IntPtr rd_kafka_incremental_assign(IntPtr rk,
+                    /* const rd_kafka_topic_partition_list_t * */ IntPtr partitions);
+
+        [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
+        internal static extern IntPtr rd_kafka_incremental_unassign(IntPtr rk,
+                      /* const rd_kafka_topic_partition_list_t * */ IntPtr partitions);
+
+        [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
+        internal static extern IntPtr rd_kafka_assignment_lost(IntPtr rk);
+
+        [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
+        internal static extern IntPtr rd_kafka_rebalance_protocol(IntPtr rk);
+
+        [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
         internal static extern ErrorCode rd_kafka_assignment(IntPtr rk,
                 /* rd_kafka_topic_partition_list_t ** */ out IntPtr topics);