+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.6.0-PRE1
+
+- References librdkafka 1.6.0. Refer to the [release notes](https://github.com/edenhill/librdkafka/releases/tag/v1.6.0) for more information. Headline features:
+  - KIP-429: Incremental rebalancing.
+  - KIP-480: Sticky partitioner.
+
+
 # 1.5.2
 
 ## Enhancements